release: prepare v0.3.0
- Bump workspace version from 0.3.0-dev to 0.3.0
- Move [Unreleased] changelog entries to [0.3.0] - 2026-03-09
- Rename planned [0.3.0] section to [0.4.0]
- Fix all clippy warnings across workspace (let-else, dead code,
  complex types, module inception, manual strip, and more)
- Fix version-hardcoded test assertions to use env!("CARGO_PKG_VERSION")
- Add #![allow(clippy::print_stdout)] to examples and benchmarks
- Run cargo fmt --all

+++ crates/ferrite-studio/src/studio/dashboard.rs
@@ -562,7 +562,7 @@ impl Dashboard {
             .unwrap_or(0.0) as u64;
         let memory_total = self
             .latest_metric_value(&metrics, "memory_total_bytes")
-            .unwrap_or(16u64 as f64 * 1024.0 * 1024.0 * 1024.0) as u64;
+            .unwrap_or(16_f64 * 1024.0 * 1024.0 * 1024.0) as u64;
         let keys_total = self
             .latest_metric_value(&metrics, "keys_total")
             .unwrap_or(0.0) as u64;
